requirements.txt
django==5.1.1
websockets==13.0.1
channels==4.1.0
daphne==4.1.2
channels_redis==4.2.0
psycopg==3.2.1
python-dotenv==1.0.1
bcrypt==3.2.0
<<<<<<< HEAD
twilio==9.3.7
selenium
=======
twilio==9.4.1
djangorestframework-simplejwt==5.3.1
djangorestframework==3.15.2
django-split-settings==1.3.2
python-dotenv==1.0.1
bcrypt==3.2.0
>>>>>>> main
